backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
python-dotenv==1.0.0
alembic==1.12.1
qdrant-client==1.7.0
openai==1.3.7

# Basic text processing
textblob==0.17.1
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10
lxml==4.9.3

# Async and performance
aiohttp==3.9.1
httpx==0.25.2

# Database and caching
redis==5.0.1
celery==5.3.4
